+++ lib/changelog/changelog_file.rb
@@ -45,16 +45,9 @@ module Danger
             @your_contribution_here = true
             next
           end
-          next if valid_line?(line)
+          next if Danger::Changelog::ChangelogLine.valid?(line)
           @bad_lines << line
         end
-      end
-
-      # match the PR format, with or without PR number
-      def valid_line?(line)
-        return true if line =~ %r{^\*\s[\`[:upper:]].* \- \[\@[\w\d\-\_]+\]\(https:\/\/github\.com\/.*[\w\d\-\_]+\)\.$}
-        return true if line =~ %r{^\*\s\[\#\d+\]\(https:\/\/github\.com\/.*\d+\)\: [\`[:upper:]].* \- \[\@[\w\d\-\_]+\]\(https:\/\/github\.com\/.*[\w\d\-\_]+\)\.$}
-        false
       end
     end
   end