+++ lib/danger_plugin.rb
@@ -1,2 +1,3 @@
+require 'changelog/changelog_line'
 require 'changelog/changelog_file'
 require 'changelog/plugin'