+++ spec/spec_helper.rb
@@ -9,7 +9,6 @@ require 'pry'
 require 'rspec'
 require 'danger'
 
-# Use coloured output, it's the best.
 RSpec.configure do |config|
   config.filter_gems_from_backtrace 'bundler'
   config.color = true
@@ -18,30 +17,10 @@ end
 
 require 'danger_plugin'
 
-# These functions are a subset of https://github.com/danger/danger/blob/master/spec/spec_helper.rb
-# If you are expanding these files, see if it's already been done ^.
-
-# A silent version of the user interface,
-# it comes with an extra function `.string` which will
-# strip all ANSI colours from the string.
-
-# rubocop:disable Lint/NestedMethodDefinition
 def testing_ui
-  @output = StringIO.new
-  def @output.winsize
-    [20, 9999]
-  end
-
-  cork = Cork::Board.new(out: @output)
-  def cork.string
-    out.string.gsub(/\e\[([;\d]+)?m/, '')
-  end
-  cork
+  Cork::Board.new(silent: true)
 end
-# rubocop:enable Lint/NestedMethodDefinition
 
-# Example environment (ENV) that would come from
-# running a PR on TravisCI
 def testing_env
   {
     'HAS_JOSH_K_SEAL_OF_APPROVAL' => 'true',