+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 0.2.0 (Next)
 
+* [#13](https://github.com/dblock/danger-changelog/pull/13): Fix: suggested CHANGELOG entry may not be of correct format - [@dblock](https://github.com/dblock).
 * Your contribution here.
 
 ### 0.1.0 (8/26/2016)